requirements.txt
Faker==8.8.1
kafka-python==2.0.2
bson==0.5.10
pymongo==3.11.4
pyspark==3.0.2